Refactor tsbasic for code simplicity
